Move recipe info to its own class
This change is required to access metadata from _recipe.py
